No ticket: [refactor] Moved the MockMediaLibrary and MockMediaPlayer classes into their own files.
